requirements.txt
flask
flask-cors
openai
pyaudio
SpeechRecognition
transformers
torch
langchain
faiss-cpu
python-dotenv
gTTS
streamlit
pytesseract
moviepy
opencv-python
numpy
langchain-openai
langchain-community
langdetect
pypdf
openpyxl
python-docx
easyocr
streamlit-mic-recorder
Pillow